+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 ## CHANGES
 
+### 16-Nov-2023
+- added function to download from releases only and not from upstream repo
+- ```bash
+chmod +x release.sh
+./release.sh
+```
+
 ### 15-Nov-2023
 - edited pywal.css waybar style.
 - waybar configs and styles renamed.